Bil test data
Add initData procedure with helper repositories and models

This commit includes the creation of an initData procedure that generates test data for the system. Also introduces BilRepository, LejeAftaleRepository, and SkadeRapportRepository to interact with the database. Additionally, the Bil model and Lejeaftale model were updated to correctly interact with the database.

Signed-off-by: AlekOmOm <142972647+AlekOmOm@users.noreply.github.com>