u1 Backend - Logge ind - Add Bruger DTO, Controller, and Service functionality
Added a User DTO class, updated the User Service with login and registration functionality, and implemented corresponding methods in the User Controller. Also, renamed Login.html to Home.html and included forms for user login and registration. Changes in the User model are primarily about enhancing data management and representation.

Signed-off-by: AlekOmOm <142972647+AlekOmOm@users.noreply.github.com>