Add explicit homepage route, improve session handling, and Home.css added
The update introduces an explicit /home routing in the controller and refactors the code to better handle sessions. It also introduces clean-up functions for session attributes and adds a service to the 'BrugerDto' model that securely handles password clearance. Minor adjustments were also made in the 'BilRepository' to correctly set the kilometer run, and a 'Home.css' file was added with new color variables and body styling rules.

Signed-off-by: AlekOmOm <142972647+AlekOmOm@users.noreply.github.com>