/Abonnement page - mapping input to nyBrugerValgDTO Controllers - tilpasset flow LejeAftale - int totalPris
The 'LejeAftale' model was simplified and an 'int totalPris' field was added. This required updating constructors, getter and setter methods, and edit on the relevant jdbcTemplate actions. HTML input form updates were made accordingly. The 'PackageDeals' model has been refactored to 'PackageDeal' and moved to the DTO package. 'name' property in html files replaced with 'packageName'.

Signed-off-by: AlekOmOm <142972647+AlekOmOm@users.noreply.github.com>