u5 testet og fuldt funktionel
SkadeRepository.findNew
- opdateret med .query til at håndtere 0 returneringer

Updated Skade model and related templates, and added debugging statements in SkadeRepository

The `toString` method was implemented for the Skade model, type of damage is now displayed instead of the name in SeSkadeRapport.html template, and the URL generation for viewing the damage report was modified in SkadeRapport.html. In SkadeRepository, debug output statements were added to the `saveAll`, `find` and `exists` methods and SELECT_ALL_WITHOUT_ID query was introduced.

Signed-off-by: AlekOmOm <142972647+AlekOmOm@users.noreply.github.com>