Enforce non-null for SkadeRapportID and refactor CPR_NR to CprNr
The SkadeRapportID field in the Skader table is updated to be non-null to improve data integrity. Additionally, the CPR_NR property in the KundeInfo class has been refactored to CprNr to comply with the Java naming conventions. This refactoring is reflected everywhere the property is used across the KundeInfoRepository class.

Signed-off-by: AlekOmOm <142972647+AlekOmOm@users.noreply.github.com>