Rename ForretningsUdviklingsController to ForretningsRapportController and add debug message in DashboardService
The ForretningsUdviklingsController java file, along with its public class and constructor, have been renamed to ForretningsRapportController. In the DashboardService java file, a debug message was added in the seTotalIndkomst method to assist with monitoring the total income. In addition, the SeSkadeRapport.html file has been reformatted to include div class properties for better page structure.

Signed-off-by: AlekOmOm <142972647+AlekOmOm@users.noreply.github.com>